pyproject.toml
[project]
name = "stlc-copilot"
version = "0.1.0"
description = ""
authors = [
    {name = "testName",email = "test@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "stlc_copilot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn src.stlc_copilot.main:app --reload"